% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl_vox_create.R
\name{ctrl_vox_create}
\alias{ctrl_vox_create}
\title{Control Voxel Create}
\usage{
ctrl_vox_create(p_map_infile, threshold = 0.99, samples = c(1500, 1200,
  1000), outfile = NULL)
}
\arguments{
\item{p_map_infile}{list or vector of size 3, for the three probabilty maps corresponding to each tissue class}

\item{threshold}{cutoff probabilty value to choose good quality tissues, default = 0.99}

\item{samples}{vector of size 3 determining the subsample sizes for White Matter, Grey Matter and CSF respectively, default =c(1500,1200,1000)}

\item{outfile}{outputs a mask of selected controls voxels to create intensity landmarks}
}
\value{
mask image of voxels where total number of 1's is sum(samples)
}
\description{
This function creates a mask to obtain ideal tissue to create landmarks for histogram matching
}
