% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landmarks.R
\name{get.landmarks}
\alias{get.landmarks}
\title{Create Landmarks}
\usage{
get.landmarks(rawdata, i.min, i.max, i.s.min, i.s.max, h, mask)
}
\arguments{
\item{mask}{this is a test}
}
\value{
A vector of landmarks to be used for Histogram Matching
}
\description{
This function creates landmark points for piecewise histogram matching
}
