% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_object.R
\name{check_object}
\alias{check_object}
\title{Check_Object function}
\usage{
check_object(object)
}
\value{
An image, either same as input or read from path
}
\description{
to verify whether NIFTI or path
}
