% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl_vox_create.R
\name{ctrl_vox_create}
\alias{ctrl_vox_create}
\title{Control Voxel Create}
\usage{
ctrl_vox_create(wm_pve, gm_pve, csf_pve, threshold = 0.99, samples = c(1500,
  1200, 1000), outfile = NULL)
}
\arguments{
\item{wm_pve}{White Matter partial volume probabilty map or path}

\item{gm_pve}{Grey Matter partial volume probabilty map or path}

\item{csf_pve}{CSF partial volume probabilty map or path}

\item{threshold}{cutoff probabilty value to choose good quality tissues, default = 0.99}

\item{samples}{vector of size 3 determining the subsample sizes for White Matter, Grey Matter and CSF respectively, default =c(1500,1200,1000)}

\item{outfile}{outputs a mask of selected controls voxels to create intensity landmarks}
}
\value{
mask image of voxels where total number of 1's is sum(samples)
}
\description{
This function creates a mask to obtain ideal tissue to create landmarks for histogram matching
}
