% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code_HM.R
\name{do.hist.norm}
\alias{do.hist.norm}
\title{Histogram Matching Piecewise Algorithm}
\usage{
do.hist.norm(rawdata, i.min, i.max, i.s.min, i.s.max, h, m, ctrl_mask,
  rangemax = 1)
}
\arguments{
\item{rawdata}{The subject image to be matched}

\item{i.min}{min for subj landmark}

\item{i.max}{max for subj landmark}

\item{i.s.min}{min for control landmark}

\item{i.s.max}{max for control landmark}

\item{h}{series or quantiles}

\item{m}{vector of landmarks obtained from control mask}

\item{ctrl_mask}{Mask created by subsampling to obtain clean landmarks}

\item{rangemax}{default = 255, Image intensities will range between 0 and 255}
}
\value{
Returns Normalized image
}
\description{
Histogram Matching Piecewise Algorithm
}
