% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landmarks.R
\name{get.landmarks}
\alias{get.landmarks}
\title{Create Landmarks}
\usage{
get.landmarks(data, i.min, i.max, i.s.min, i.s.max, h, mask)
}
\arguments{
\item{data}{image}

\item{i.min}{min for subj landmark}

\item{i.max}{max for subj landmark}

\item{i.s.min}{min for control landmark}

\item{i.s.max}{max for control landmark}

\item{h}{series or quantiles}

\item{mask}{this is corresponding control mask obatined from ctrl_vox_create}
}
\value{
A vector of landmarks to be used for Histogram Matching
}
\description{
This function creates landmark points for piecewise histogram matching
}
