% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatch.R
\name{hatch}
\alias{hatch}
\title{hatch}
\usage{
hatch(infile, prob_maps_infile, ctrl_vox, ctrl_img, i.min = 0.01,
  i.max = 0.99, i.s.min = 0, i.s.max = 1, h = seq(0.1, 0.9, by = 0.1),
  rangemax = 255, threshold = 0.97, samples = c(1500, 1200, 1000),
  outfile = NULL)
}
\arguments{
\item{infile}{Patient Image Nifti file  or path (T1, T2, T1gad or FLAIR)}

\item{prob_maps_infile}{List or vector of 3 partial volume probability maps paths associated with the three different types of Tissue Classes or a list of the 3 nifti files}

\item{ctrl_vox}{Control Mask created using healthy control of subsampled voxles}

\item{ctrl_img}{Healthy Control image coreesponding to the modality of the subject}

\item{h}{series or quantiles}

\item{threshold}{cutoff probabilty value to choose good quality tissues, default = 0.99}

\item{samples}{vector of size 3 determining the subsample sizes for White Matter, Grey Matter and CSF respectively, default =c(1500,1200,1000)}

\item{outfile}{path to save output matched class to a particular location}
}
\value{
The histogram matched normalized image
}
\description{
Takes in subject input file and healthy control files to produce matched intensity histograms
}
